WIP: more query db stuff
