+++ crates/nu-command/src/database/mod.rs
@@ -1,3 +1,3 @@
 mod sqlite;
 
-pub use sqlite::SQLiteConnection;
+pub use sqlite::{SQLiteDatabase, read_sqlite_db, open_and_read_sqlite_db};
